Seperate controller code from trait code. This makes it easier for people to reuse the code from the trait and implement their own JsonApiController. If one needs to build a supplemented resource controller for example, he can not inherit from the JsonApiController provided by this package because the methods inside a supplemented resource controller will have other signatures (more paramaters)
