+++ src/NilPortugues/Laravel5/JsonApi/Controller/JsonApiTrait.php
@@ -13,8 +13,6 @@ namespace NilPortugues\Laravel5\JsonApi\Controller;
 use Illuminate\Container\Container;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Http\Request;
-use NilPortugues\Api\JsonApi\Http\Factory\RequestFactory;
-use NilPortugues\Api\JsonApi\Http\Response\ResourceNotFound;
 use NilPortugues\Laravel5\JsonApi\Actions\CreateResource;
 use NilPortugues\Laravel5\JsonApi\Actions\DeleteResource;
 use NilPortugues\Laravel5\JsonApi\Actions\GetResource;
@@ -45,36 +43,6 @@ trait JsonApiTrait
     public function __construct(JsonApiSerializer $serializer)
     {
         $this->serializer = $serializer;
-    }
-
-    /**
-     * Get many resources.
-     *
-     * @return \Symfony\Component\HttpFoundation\Response
-     */
-    public function index()
-    {
-        $apiRequest = RequestFactory::create();
-
-        $page = $apiRequest->getPage();
-        if (!$page->size()) {
-            $page->setSize($this->pageSize);
-        }
-
-        $fields = $apiRequest->getFields();
-        $sorting = $apiRequest->getSort();
-        $included = $apiRequest->getIncludedRelationships();
-        $filters = $apiRequest->getFilters();
-
-        $resource = new ListResource($this->serializer, $page, $fields, $sorting, $included, $filters);
-
-        $totalAmount = $this->totalAmountResourceCallable();
-        $results = $this->listResourceCallable();
-
-        $controllerAction = '\\'.get_called_class().'@index';
-        $uri = $this->uriGenerator($controllerAction);
-
-        return $this->addHeaders($resource->get($totalAmount, $results, $uri, get_class($this->getDataModel())));
     }
 
     /**
@@ -133,36 +101,6 @@ trait JsonApiTrait
     }
 
     /**
-     * @return ResourceNotFound
-     */
-    public function create()
-    {
-        return new ResourceNotFound();
-    }
-
-    /**
-     * Get single resource.
-     *
-     * @param $id
-     *
-     * @return \Symfony\Component\HttpFoundation\Response
-     */
-    public function show($id)
-    {
-        $apiRequest = RequestFactory::create();
-
-        $resource = new GetResource(
-            $this->serializer,
-            $apiRequest->getFields(),
-            $apiRequest->getIncludedRelationships()
-        );
-
-        $find = $this->findResourceCallable($id);
-
-        return $this->addHeaders($resource->get($id, get_class($this->getDataModel()), $find));
-    }
-
-    /**
      * @param $id
      *
      * @return callable
@@ -176,23 +114,6 @@ trait JsonApiTrait
 
             return $model;
         };
-    }
-
-    /**
-     * Post Action.
-     *
-     * @param Request $request
-     *
-     * @return \Symfony\Component\HttpFoundation\Response
-     */
-    public function store(Request $request)
-    {
-        $createResource = $this->createResourceCallable();
-        $resource = new CreateResource($this->serializer);
-
-        return $this->addHeaders(
-            $resource->get((array) $request->get('data'), get_class($this->getDataModel()), $createResource)
-        );
     }
 
     /**
@@ -223,17 +144,6 @@ trait JsonApiTrait
 
             return $model;
         };
-    }
-
-    /**
-     * @param $id
-     *
-     * @return Response
-     */
-    public function update(Request $request, $id)
-    {
-        return (strtoupper($request->getMethod()) === 'PUT') ? $this->putAction($request,
-            $id) : $this->patchAction($request, $id);
     }
 
     /**
@@ -300,30 +210,6 @@ trait JsonApiTrait
                 $update
             )
         );
-    }
-
-    /**
-     * @return ResourceNotFound
-     */
-    public function edit()
-    {
-        return new ResourceNotFound();
-    }
-
-    /**
-     * @param $id
-     *
-     * @return Response
-     */
-    public function destroy($id)
-    {
-        $find = $this->findResourceCallable($id);
-
-        $delete = $this->deleteResourceCallable($id);
-
-        $resource = new DeleteResource($this->serializer);
-
-        return $this->addHeaders($resource->get($id, get_class($this->getDataModel()), $find, $delete));
     }
 
     /**
